outsource some functions to yeeter.js to other files
