+++ index.html
@@ -78,7 +78,7 @@
       </div>
     </div>
     <br />
-    <button id="open-settings" class="btn">toggle advanced settings menu</button>
+    <button id="open-settings-btn" class="btn">toggle advanced settings menu</button>
     <br />
     <div>
       <br />
@@ -108,7 +108,7 @@ Kahoot Yeeter program started.
     </div>
 
     <div id="sidenav" class="sidenav">
-      <a href="javascript:void(0)" class="closebtn" id="close-nav">&times;</a>
+      <a href="javascript:void(0)" class="closebtn" id="close-nav-btn">&times;</a>
       <a id="quiz-name" href="#">unknown</a>
       <a id="question-amount" href="#">unknown</a>
       <a id="current-question" href="#">unknown</a>
